Fix Arduino/Genuino 101 compatibility

+++ examples/USBH_MIDI_dump/USBH_MIDI_dump.ino
@@ -1,7 +1,7 @@
 /*
  *******************************************************************************
  * USB-MIDI dump utility
- * Copyright 2013-2016 Yuuichi Akagawa
+ * Copyright 2013-2017 Yuuichi Akagawa
  *
  * for use with USB Host Shield 2.0 from Circuitsathome.com
  * https://github.com/felis/USB_Host_Shield_2.0
@@ -37,7 +37,7 @@ USB Usb;
 USBH_MIDI  Midi(&Usb);
 
 void MIDI_poll();
-void doDelay(unsigned long t1, unsigned long t2, unsigned long delayTime);
+void doDelay(uint32_t t1, uint32_t t2, uint32_t delayTime);
 
 boolean bFirst;
 uint16_t pid, vid;
@@ -60,16 +60,14 @@ void setup()
 
 void loop()
 {
-  //unsigned long t1;
-
   Usb.Task();
-  //t1 = micros();
+  //uint32_t t1 = (uint32_t)micros();
   if ( Usb.getUsbTaskState() == USB_STATE_RUNNING )
   {
     MIDI_poll();
   }
   //delay(1ms)
-  //doDelay(t1, micros(), 1000);
+  //doDelay(t1, (uint32_t)micros(), 1000);
 }
 
 // Poll USB MIDI Controler and send to serial MIDI
@@ -86,7 +84,7 @@ void MIDI_poll()
     pid = Midi.pid;
   }
   if (Midi.RecvData( &rcvd,  bufMidi) == 0 ) {
-    sprintf(buf, "%08lX: ", millis());
+    sprintf(buf, "%08lX: ", (uint32_t)millis());
     Serial.print(buf);
     Serial.print(rcvd);
     Serial.print(':');
@@ -99,9 +97,9 @@ void MIDI_poll()
 }
 
 // Delay time (max 16383 us)
-void doDelay(unsigned long t1, unsigned long t2, unsigned long delayTime)
+void doDelay(uint32_t t1, uint32_t t2, uint32_t delayTime)
 {
-  unsigned long t3;
+  uint32_t t3;
 
   if ( t1 > t2 ) {
     t3 = (0xFFFFFFFF - t1 + t2);
